Made DREAM initialize without any errors
This commit introduces several new classes, as well as some changes to
already existing classes. Most notably, a new set of EquationTerm's have
been added:

- PredeterminedQuantity
  - ConstantParameter
  - PrescribedParameter
- MomentQuantity

The 'PredeterminedQuantity' replaces 'PrescribedParameter' as the base
class for terms that shoule be evaluated prior to solving the equation
system, i.e. for terms which should not show up in the equation system
matrix. The 'MomentQuantity' represents a quantity that is determined as
the moment of another unknown quantity in the equation system (such as,
for example, the hot electron density, which is the first moment of the
hot electron distribution function).

Several equation terms have also been modified slightly to use the
'UnknownQuantityHandler' instead of the  full 'EquationSystem' class for
getting unknown data.

A few changes had to be made to the 'CollisionQuantityHandler' in order
to make it compile (the linker had previously discarded that object as
it was never used anywhere in the code). I have implemented a few
missing deallocators as empty methods and added pre-processor warnings
to their bodies.

Finally, this commit makes it possible to run DREAM with the 'basic'
example 'dream_settings.h5' file without any errors or memory errors
(checked with valgrind). Note, though, that this just means that the
simulation is initialized to some extent, but no actual solution is
attempted.