Fixed RHS bug in linear solver
This bug was most notably observed in the prescribed ion densities,
which turned out wrong. The reason for this was that, in
'Solver::BuildMatrix()', the row/col offset was only set on the matrix,
and not on the RHS vector. Since we haven't looked too much at equations
with non-zero RHS vectors this has gone under the radar.