Flipped sign of RHS in linearized equations.
Previously, we solved equations of the form

  Mx = S

which is exactly how PETSc does it. Now, we instead solve equations of
the form

  Mx + S = 0

which requires us to negate the vector 'S' before handing it to PETSc.
This sign change has been made to be consistent with the nonlinear
solver, where we solve F(x) = 0 with

  F ~ Mx + S

in its linearized form. (More explicitly, this formulation allows us to
use 'SetVectorElements()', which is intended for building the non-linear
function F(x), to insert the values of trivial quantities into the RHS
in the linear solver).

We also fixed a number of related sign-bugs in this commit. Non-linear
solver remains non-functional.