Restructured 'EvaluableEquationTerm'.
Now, the 'Evaluable' property of an EquationTerm means that it has the
method 'EvaluableTransform()' which transforms a vector so that it
contains the value of the unknown quantity to which the EquationTerm is
applied.

For example, if the EquationTerm represents the first term in the
equation

  f(x) + g(y) = 0,

where x is the unknown quantity, f(x) is an invertible function of only
x, y is any other unknown quantity (or set of unknown quantities) and
g(y) is an arbitrary function of those. The 'EvaluableTransform()' then
takes a vector 'vec' as input with elements initialized to

  vec[i] = g_i(y).

On return, the vector has been transformed to instead contain

  vec[i] = f^-1( -g_i(y) ),

i.e. x; where f^-1(x) denotes the inverse of f(x).