Update PXiExternalLoss boundary condition
The 'AddToMatrixElements()' and 'AddToVectorElements()' were combined
and support was added for two new boundary conditions (f=0 at p_{N_p+1}
and dPhi/dp_{N_p+1/2} = dPhi/dp_{N_p-1/2} on the boundary.

The 'PXiExternalLoss' boundary condition can now also be applied to
fluid quantities to imitate a source term. This means that
'PXiExternalLoss' now replaces 'DensityFromBoundaryFlux' which was
previously used to this end. It further means that the boundary
condition is now implemented in just one location the code
(PXiExternalLoss.cpp) instead of four independent locations as before.
The old 'DensityFromBoundaryFlux' is nonetheless kept for testing
purposes, as it is used by the 'BoundaryFlux' test in 'dreamtests'.

Note that the "f=0" boundary condition requires knowledge of
   dp_{Np} = p_{N_p+1} - p_{N_p},
which is an arbitrary number. Currently, it is fixed to be dp_{N_p-1/2}.