Implemented 'DurationTimer'
The duration timer can be used accumulate time measurements of sections
of the code. As a (useful) example, I have integrated a few
DurationTimer's into the SolverLinearlyImplicit and SolverNonLinear
objects in order to measure the percentages of time spent rebuilding
coefficients, residuals, matrices, jacobians and for inverting matrices.
This approach, although more superficial than, for example, callgrind,
will be able to provide accurate measurements of simulation times even
when compiling with '-O3' (highest optimization).

The DurationTimer uses the C++ std::chrono library and a
'std::chrono::high_resolution_clock' to measure performance. Typically
this clock is able to measure time very accurately (on a microsecond
scale).