Fixed bugs in kinetic/kinetic boundary condition
The test still gives errors when comparing fluxes according to equation
(10) in 'doc/notes/discretisation.pdf', i.e. when comparing fluxes
element-wise for grids with mismatching xi grids (i.e. when nxi_hot and
nxi_RE are not multiples). Since the total number of particles is always
conserved, however, this may be the correct behaviour, suggesting a
minor logical error my notes. This error should luckily not affect
particle conservation or physics, but it might be worth further
investigating this behaviour in the future.

This means that the boundary condition should work for advection
problems. Next up is to implement support for diffusion (which will
complete the implementation), but this first requires a separate (but
similar (to the advection term)) derivation.