Implemented support for runaway distribution
Seems to work well when nxi_hot = nxi_RE, but fails otherwise. The
failing behaviour is similar to what happens when an advection term is
applied without any stabilizing diffusion. We should hence check whether
the problem is in fact due to the boundary condition (which passes a
fairly sophisticated test) or if it's due to something else (a mismatch
with coefficients on the runaway grid? missing terms on the runaway
grid?)