Ported class 'ConvergenceScan' from NORSE++
The purpose of this class is to make running convergence scans much
easier. An example is provided for the 'runaway' example under
"examples/convergence", which sets up the same 'DREAMSettings' object as
in "examples/runaway", and then passes it to a 'ConvergenceScan' for
scanning in the resolution parameters 'np', 'nxi' and 'nt'.