Add support for calculating/plotting normalized E
in Python interface. All plot functions associated with E_field now have
a 'norm' argument which specifies which quantity to normalize the
electric field by (key into 'DREAMOutput.other.fluid'), as well as the
'normalize()' method (along with its alias 'norm()') which constructs a
new FluidQuantity which is a normalized electric field.