Adaptive time stepper works
It seems that the adaptive time stepper works. The 'ts_adaptive' test
(tests/physics/ts_adaptive), which overrides the "adaptive" stepper and
forces a constant time step (which allows exact comparison to constant
time stepper and verification that rolling back to previous solutions
doesn't affect the final solution) passes indicating that any remaining
errors should be contained in the determination of the new time step.
The 'ts_adaptive' example behaves mostly well, except in the end where
it overshoots the final time by quite a bit. This should be solved
before finally merging to master.