Fixed bug occuring for older h5py
In a previous commit we switched to storing strings with UTF-8 encoding
in HDF5 files. This was apparently only supported since h5py version
2.10.0, which has only been available since September 2019 (far too
recently for the package to be available on Ubuntu...), and so the
interface is now causing some problems on systems with older versions of
h5py. This commit should fix the problem on older systems while
retaining UTF-8 support on systems with more recent versions of h5py.