Added code for saving PETSc citations
PETSc has a command-line flag "-citations" for storing the citations
appropriate for a given PETSc run to be saved to the specified text file
after the run (in BibTeX format). This commit adds the necessary code
for achieving this in the dreami interface (it's not a very pretty
solution, but most likely we will rarely need the functionality).