Prepared for using backup inverter
Added a "backup inverter" that can be used when the primary linear
solver (inverter) fails to converge. This is due to observations of
the PETSc LU solver converging in some cases where MUMPS does not. The
backup solver should thus be applied when the primary solver does not
converge.

The implementation is mostly finished, and the only part lacking is the
ability of the TimeStepperConstant to roll back the solution.