Moved the calculation of pBarInv to separate function in the template class
On branch SvenssonTransport
Your branch is up to date with 'origin/SvenssonTransport'.

Changes to be committed:
	modified:   include/DREAM/Equations/Fluid/SvenssonTransport.hpp
	modified:   include/DREAM/Equations/Fluid/SvenssonTransport.tcc
	modified:   src/Equations/Fluid/SvenssonTransport.cpp

The now separate function has provisions for also calculating the
derivative of pBarInv with respect to r. However, this provision can
perhaps beoptimized a bit for the cases where the dervative isn't
needed.

The code compiles on my office computer.