SvenssonTransport is now incorporated in TransportSettings.
On branch SvenssonTransport
Your branch is up to date with 'origin/SvenssonTransport'.

Changes to be committed:
	modified:   include/DREAM/Equations/Fluid/SvenssonTransport.hpp
	modified:   include/DREAM/Equations/Fluid/SvenssonTransport.tcc
	modified:   src/Equations/Fluid/SvenssonTransport.cpp
	modified:   src/Settings/Transport.cpp

The incorporation of SvenssonTransport in TransportSettings has been
done without a timedependence.

The pitch-angle is not fully implemented. So far it is a simple for
loop sum, not a GSL integral.

To do in the future
 - Time dependance should be added (back).
 - Pitch-angle averaging with a GSL integral.
 - Pitch-angle averaging with the prescribed distribution.