Added timedependance to the coefficients for SvenssonTransport.
On branch SvenssonTransport
Your branch is up to date with 'origin/SvenssonTransport'.

Changes to be committed:
	modified:   examples/SvenssonTransport/generate.py
	modified:   include/DREAM/Equations/Fluid/SvenssonTransport.hpp
	modified:   include/DREAM/Equations/Fluid/SvenssonTransport.tcc
	modified:   py/DREAM/Settings/TransportSettings.py
	modified:   src/Equations/Fluid/SvenssonTransport.cpp
	modified:   src/Settings/Transport.cpp

Time dependent input is done via the dream_4d_data struct, and then
(nearest-index) interpolating the data in time.

Started to implement pitch-angle averaging. Since the prescribed
distribution has a time dependence, the pitch-angle averaging will
have to be done at every time step.

**Still to do**

At the moment, the pstar lower bound is not implemented, which means
that if the input data goes down to p=0, the pitch-angle averaging
will generate NaN and crach the simulation.

If the input data is given on ppar-pperp, soething should be
done. Either to convert the input grid to p-xi, or to do the
integrations in ppar-pperp. My guess, for now, ist to convert to p-xi.