Some (pre-testing) optimization work.
On branch SvenssonTransport
Your branch is up to date with 'origin/SvenssonTransport'.

Changes to be committed:
	modified:   examples/SvenssonTransport/generate.py
	modified:   include/DREAM/Equations/Fluid/SvenssonTransport.tcc
	modified:   src/Equations/Fluid/SvenssonTransport.cpp

Optimized the xi average quite a bit by realizing that the pitch-angle
distribution width can be calculated outside the xi loop.

Also went over the code and did some cleanup and added comments.

**Still remaining** One should probably add the time interpolation
method as a setting to be read in from the settings file. At the
moment, that option is hardcoded as INTERP1D_NEAREST.