Changed initialization of f_re
Now we're using EqsysInitializer instead (a somewhat more complicated,
but also nicer, solution) of manually loading the prescribed n_re
multiple times. This also allowed me to put all of n_re in the xi cell
(+1 or -1) corresponding to the "forward" direction of the distribution,
i.e. the direction in which the electric field will accelerate
particles.