Implemented workaround for memory bug
Previously, to avoid PETSc reallocating memory for some non-zero
elements in the second iteration of the Newton solver, we put in a flag
which only ignored explicitly zero values in the Matrix after the very
first call to 'Assemble()' (which happens in connection with the first
Newton iteration). It however seems that the linear solvers were not
able to distinguish between explicit zeros and non-zero elements in the
matrix, and so they would run extremely slowly (and sometimes even
request huge amounts of memory).

As a workaround, we now instead delete the jacobian matrix after first
iteration and reconstruct it from scratch, effectively resetting the
non-zero structure. This takes a bit of time, but is significantly
cheaper than letting PETSc allocate memory for non-zeros which were not
set in the first iteration. It also greatly speeds up the subsequent
linear solves.