Make sure to rebuild IonHandler in the initializer
Prior to this commit, the IonHandler was only re-built when the ions
were initialized in the "usual" way, while not at all when the ions were
to be loaded from a DREAM output file. In previous versions of DREAM,
this has probably gone undetected due to that there used to be a call to
'IonHandler::Rebuild()' in 'src/Settings/Equations/ions.cpp' which was
always executed. However, probably many continuation simulations may
therefore have been initialized with the wrong densities (which would
have been corrected after the first time step).