Do not return zero if ion time scale is infinite
Otherwise, if you have a very long ion time scale in the beginning of the simulation (such as for SPI simulations) you will have to run with the initial time step for very long. This is off course not a huge problem if the initial time step is not set too short, but I anyway don't see the point of not allowing infinite ion time scales.