Corrected analytical Svensson transport coefficient set by TransportCoefficientReader
The xi-grid for the analytical diffusion coefficient now only consists of one point, so that the diffusion coefficient is taken at face value, to avoid the problems described in issues 414 and 415. The resolution on the p-grid on which the diffusion coefficient is provided is also increased, to increase the accuracy of the p average.